split_columns should be part of the List module
